Resolves the eight conflicts main's restructuring opened: the build
command's option list and refusal helper (my refusals now go through
`_refuse` with registered codes), the master-data request check (the
`identifiers` switch beside main's vocabulary limits), the wrap-tolerant
assertion both sides made, and the four rewritten documents, onto which
the seams' insertions are re-applied.

Adapts to gates main added since the branch was cut: `calibrate` draws
seeded noise through `worldloom.rng` rather than importing `random`
(the determinism AST gate allows one such module, and it is right to);
`fidelity` sorts its one set iteration; causal traces subclass the thin
waist's model so `World.causal` is a bounded `Collection`; ruff and mypy
are clean; `causal` is named in `_install()`; the surface rules gain
Saudi Arabia and Kuwait, which the Gulf locale now spans.

+++ src/worldloom/providers.py
@@ -359,7 +359,7 @@ def _allocate(total: float, weights: Sequence[float], decimals: int) -> list[flo
     the column sums to the total exactly at that precision by construction.
     """
     scale = 10 ** decimals
-    units = int(round(total * scale))
+    units = round(total * scale)
     weight_sum = sum(weights) or 1.0
     exact = [units * w / weight_sum for w in weights]
     floors = [int(x) for x in exact]
